# docs/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/auth/password/register:
    post:
      description: Creates a user and immediately issues first-party access + refresh
        tokens.
      operationId: auth.password.register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordRegisterRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResultWithMeEnvelopeDto"
      summary: Register (password)
      tags: &a1
        - Auth
      x-error-codes:
        - VALIDATION_FAILED
        - AUTH_EMAIL_ALREADY_EXISTS
        - INTERNAL
  /v1/auth/oidc/exchange:
    post:
      description: Verifies an OIDC id_token (e.g., Google) and issues first-party
        access + refresh tokens. Does not auto-link to an existing password
        account purely by email.
      operationId: auth.oidc.exchange
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OidcExchangeRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResultWithMeEnvelopeDto"
      summary: Exchange OIDC id_token
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - AUTH_OIDC_NOT_CONFIGURED
        - AUTH_OIDC_TOKEN_INVALID
        - AUTH_OIDC_EMAIL_NOT_VERIFIED
        - AUTH_OIDC_LINK_REQUIRED
        - AUTH_USER_SUSPENDED
        - INTERNAL
  /v1/auth/oidc/connect:
    post:
      description: Links an OIDC identity (e.g., Google) to the authenticated user. If
        the OIDC email matches the user email, the user email is marked
        verified.
      operationId: auth.oidc.connect
      parameters:
        - name: Idempotency-Key
          in: header
          description: "Idempotency key for safe retries of write requests. Replays return
            `Idempotency-Replayed: true`."
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
            description: "Opaque string (recommended: UUIDv4)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OidcConnectRequestDto"
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Connect OIDC identity (current user)
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - IDEMPOTENCY_IN_PROGRESS
        - CONFLICT
        - AUTH_OIDC_NOT_CONFIGURED
        - AUTH_OIDC_TOKEN_INVALID
        - AUTH_OIDC_EMAIL_NOT_VERIFIED
        - AUTH_OIDC_IDENTITY_ALREADY_LINKED
        - AUTH_OIDC_PROVIDER_ALREADY_LINKED
        - INTERNAL
  /v1/auth/email/verify:
    post:
      description: Verifies a user email using a token sent via email.
      operationId: auth.email.verify
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequestDto"
      responses:
        "204":
          description: ""
      summary: Verify email
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - AUTH_EMAIL_VERIFICATION_TOKEN_INVALID
        - AUTH_EMAIL_VERIFICATION_TOKEN_EXPIRED
        - INTERNAL
  /v1/auth/email/verification/resend:
    post:
      description: Enqueues a new verification email for the authenticated user (rate
        limited).
      operationId: auth.email.verification.resend
      parameters: []
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Resend verification email (current user)
      tags: *a1
      x-error-codes:
        - UNAUTHORIZED
        - RATE_LIMITED
        - INTERNAL
  /v1/auth/password/reset/request:
    post:
      description: Enqueues a password reset email for an existing user. Returns 204
        even if the email is unknown to avoid account enumeration.
      operationId: auth.password.reset.request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestDto"
      responses:
        "204":
          description: ""
      summary: Request password reset
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - RATE_LIMITED
        - INTERNAL
  /v1/auth/password/reset/confirm:
    post:
      description: Resets the user password using a one-time token and revokes all sessions.
      operationId: auth.password.reset.confirm
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequestDto"
      responses:
        "204":
          description: ""
      summary: Confirm password reset
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - AUTH_PASSWORD_RESET_TOKEN_INVALID
        - AUTH_PASSWORD_RESET_TOKEN_EXPIRED
        - INTERNAL
  /v1/auth/password/login:
    post:
      description: Authenticates a user and issues first-party access + refresh tokens.
      operationId: auth.password.login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordLoginRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResultWithMeEnvelopeDto"
      summary: Login (password)
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - AUTH_INVALID_CREDENTIALS
        - AUTH_USER_SUSPENDED
        - RATE_LIMITED
        - INTERNAL
  /v1/auth/password/change:
    post:
      description: Changes the authenticated user password. Revokes other sessions
        (and their refresh tokens) but keeps the current session active.
      operationId: auth.password.change
      parameters:
        - name: Idempotency-Key
          in: header
          description: "Idempotency key for safe retries of write requests. Replays return
            `Idempotency-Replayed: true`."
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
            description: "Opaque string (recommended: UUIDv4)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequestDto"
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Change password (current user)
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - IDEMPOTENCY_IN_PROGRESS
        - CONFLICT
        - AUTH_PASSWORD_NOT_SET
        - AUTH_CURRENT_PASSWORD_INVALID
        - INTERNAL
  /v1/auth/refresh:
    post:
      description: Rotates the refresh token and returns a new access + refresh token.
      operationId: auth.refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResultEnvelopeDto"
      summary: Refresh tokens
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - AUTH_REFRESH_TOKEN_INVALID
        - AUTH_REFRESH_TOKEN_EXPIRED
        - AUTH_REFRESH_TOKEN_REUSED
        - AUTH_SESSION_REVOKED
        - AUTH_USER_SUSPENDED
        - INTERNAL
  /v1/auth/logout:
    post:
      description: Revokes the session associated with the provided refresh token.
      operationId: auth.logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequestDto"
      responses:
        "204":
          description: ""
      summary: Logout
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - AUTH_REFRESH_TOKEN_INVALID
        - INTERNAL
...
  /v1/me/account-deletion/request:
    post:
      description: Schedules account deletion 30 days in the future. The account
        remains usable during the grace period and the request can be canceled.
        After the grace period, the account is de-identified (PII erased) and
        the email becomes reusable.
      operationId: users.me.accountDeletion.request
      parameters:
        - name: Idempotency-Key
          in: header
          description: "Idempotency key for safe retries of write requests. Replays return
            `Idempotency-Replayed: true`."
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
            description: "Opaque string (recommended: UUIDv4)."
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Request account deletion (30-day grace)
      tags: &a6
        - Users
      x-error-codes:
        - UNAUTHORIZED
        - IDEMPOTENCY_IN_PROGRESS
        - CONFLICT
        - USERS_CANNOT_DELETE_LAST_ADMIN
        - INTERNAL
...
info:
  title: Backend Core Kit API
  description: Generated API contract (code-first).
  version: 0.1.0
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: "First-party access token (Authorization: Bearer <token>)"
  schemas:
    PasswordRegisterRequestDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          minLength: 10
        deviceId:
          type: string
          description: Stable per-device identifier (recommended).
        deviceName:
          type: string
          description: Human-friendly device name (optional).
      required:
        - email
        - password
    MeProfileDto:
      type: object
      properties:
        profileImageFileId:
          type: string
          example: 3d2c7b2a-2dd6-46a5-8f8e-3b5de8a5b0f0
          nullable: true
          description: Current profile image stored file ID. Use `GET
            /v1/me/profile-image/url` to get a short-lived URL for rendering.
        displayName:
          type: string
          example: Dante
          nullable: true
        givenName:
          type: string
          example: Dante
          nullable: true
        familyName:
          type: string
          example: Alighieri
          nullable: true
    AccountDeletionDto:
      type: object
      properties:
        requestedAt:
          type: string
          example: 2026-01-10T12:34:56.789Z
          format: date-time
        scheduledFor:
          type: string
          example: 2026-02-09T12:34:56.789Z
          format: date-time
      required:
        - requestedAt
        - scheduledFor
    MeDto:
      type: object
      properties:
        id:
          type: string
          example: 3d2c7b2a-2dd6-46a5-8f8e-3b5de8a5b0f0
        email:
          type: string
          example: user@example.com
        emailVerified:
          type: boolean
          example: false
        roles:
          example:
            - USER
          type: array
          items:
            type: string
        authMethods:
          type: array
          example:
            - PASSWORD
          description: Linked authentication methods on this account.
          items:
            type: string
            enum: &a12
              - PASSWORD
              - GOOGLE
        profile:
          $ref: "#/components/schemas/MeProfileDto"
        accountDeletion:
          nullable: true
          description: When set, the account is scheduled for deletion. The request can be
            canceled until scheduledFor.
          type: object
          allOf:
            - $ref: "#/components/schemas/AccountDeletionDto"
      required:
        - id
        - email
        - emailVerified
        - roles
        - authMethods
        - profile
    AuthResultWithMeDto:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/MeDto"
        accessToken:
          type: string
          example: <access-token>
        refreshToken:
          type: string
          example: <refresh-token>
      required:
        - user
        - accessToken
        - refreshToken
    AuthResultWithMeEnvelopeDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AuthResultWithMeDto"
      required:
        - data
    OidcExchangeRequestDto:
      type: object
      properties:
        provider:
          type: string
          enum: &a11
            - GOOGLE
          example: GOOGLE
        idToken:
          type: string
          example: <oidc-id-token>
        deviceId:
          type: string
          description: Stable per-device identifier (recommended).
        deviceName:
          type: string
          description: Human-friendly device name (optional).
      required:
        - provider
        - idToken
    OidcConnectRequestDto:
      type: object
      properties:
        provider:
          type: string
          enum: *a11
          example: GOOGLE
        idToken:
          type: string
          example: <oidc-id-token>
      required:
        - provider
        - idToken
    VerifyEmailRequestDto:
      type: object
      properties:
        token:
          type: string
          example: <verification-token>
      required:
        - token
    PasswordResetRequestDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
      required:
        - email
    PasswordResetConfirmRequestDto:
      type: object
      properties:
        token:
          type: string
          example: <password-reset-token>
        newPassword:
          type: string
          minLength: 10
      required:
        - token
        - newPassword
    PasswordLoginRequestDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          minLength: 1
        deviceId:
          type: string
          description: Stable per-device identifier (recommended).
        deviceName:
          type: string
          description: Human-friendly device name (optional).
      required:
        - email
        - password
    ChangePasswordRequestDto:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 10
      required:
        - currentPassword
        - newPassword
    RefreshRequestDto:
      type: object
      properties:
        refreshToken:
          type: string
          example: <refresh-token>
      required:
        - refreshToken
    AuthUserDto:
      type: object
      properties:
        id:
          type: string
          example: 3d2c7b2a-2dd6-46a5-8f8e-3b5de8a5b0f0
        email:
          type: string
          example: user@example.com
        emailVerified:
          type: boolean
          example: false
        authMethods:
          type: array
          example:
            - PASSWORD
            - GOOGLE
          description: Linked authentication methods on this account. Omitted on token
            refresh responses.
          items:
            type: string
            enum: *a12
      required:
        - id
        - email
        - emailVerified
    AuthResultDto:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUserDto"
        accessToken:
          type: string
          example: <access-token>
        refreshToken:
          type: string
          example: <refresh-token>
      required:
        - user
        - accessToken
        - refreshToken
    AuthResultEnvelopeDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AuthResultDto"
      required:
        - data
    LogoutRequestDto:
      type: object
      properties:
        refreshToken:
          type: string
          example: <refresh-token>
      required:
        - refreshToken
...